Refactor config using table instead of yaml file

+++ src/ElasticSearchBundle/Installer.php
@@ -16,6 +16,8 @@ class Installer extends MigrationInstaller
 {
     const TABLE_INDEX_RULE = 'bundle_elasticsearch_index_rule';
 
+    const TABLE_CONFIG = 'bundle_elasticsearch_config';
+
     /**
      * Executes install migration. Used during installation for initial creation of database tables and other data
      * structures (e.g. pimcore classes). The version object is the version object which can be used to add raw SQL
@@ -29,6 +31,7 @@ class Installer extends MigrationInstaller
     public function migrateInstall(Schema $schema, Version $version)
     {
         $this->installIndexRuleTable($schema, $version);
+        $this->installConfigTable($schema, $version);
     }
 
     /**
@@ -40,6 +43,7 @@ class Installer extends MigrationInstaller
     public function migrateUninstall(Schema $schema, Version $version)
     {
         $this->uninstallIndexRuleTable($schema, $version);
+        $this->uninstallConfigTable($schema, $version);
     }
 
     private function installIndexRuleTable(Schema $schema, Version $version)
@@ -91,6 +95,55 @@ class Installer extends MigrationInstaller
 
     private function uninstallIndexRuleTable(Schema $schema, Version $version)
     {
+        if (!$schema->hasTable(self::TABLE_INDEX_RULE)) {
+            return;
+        }
+
         $schema->dropTable(self::TABLE_INDEX_RULE);
+    }
+
+    private function installConfigTable(Schema $schema, Version $version)
+    {
+        $table = $schema->createTable(self::TABLE_CONFIG);
+
+        $table->addColumn('id', 'integer', [
+            'autoincrement' => true,
+        ]);
+
+        $table->addColumn('name', 'string', [
+            'notnull' => false
+        ]);
+
+        $table->addColumn('hostorip', 'string', [
+            'notnull' => false
+        ]);
+
+        $table->addColumn('port', 'string', [
+            'notnull' => false
+        ]);
+
+        $table->addColumn('httpBasicAuthUser', 'string', [
+            'notnull' => false
+        ]);
+
+        $table->addColumn('httpBasicAuthPassword', 'string', [
+            'notnull' => false
+        ]);
+
+        $table->addColumn('index', 'string', [
+            'notnull' => false
+        ]);
+
+        $table->addUniqueIndex(['name']);
+        $table->setPrimaryKey(['id']);
+    }
+
+    private function uninstallConfigTable(Schema $schema, Version $version)
+    {
+        if (!$schema->hasTable(self::TABLE_CONFIG)) {
+            return;
+        }
+
+        $schema->dropTable(self::TABLE_CONFIG);
     }
 }
